pyproject.toml
[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[project]
name = "pandas_openlineage"
version = "0.1.0"
description = "An attempt to integrate OpenLineage into Pandas"
authors = [
    { name = "Jakub Dardzinski" }
]
dependencies = [
    "pandas",
    "openlineage-python",
    "fsspec"
]
requires-python = ">=3.8, <3.13"

[project.optional-dependencies]
s3 = ["s3fs"]

[tool.setuptools.packages.find]
include = ["pandas_openlineage.*"]
